1. Arduino의 Wire의 buffer가 32이므로 I2C_TRANSFER_SIZE를 24로 변경함.
2. beseflight 프로젝트에 있던 HMC5883 보정 기능을 PIF로 옮겨 삭제함.pif_Delay1ms
3. Sensor의 Init 함수에 Param 기능 추가에 따른 수정.
4. IMU 센서값을 int16_t에서 float형으로 변경함.
5. I2C 센서의 Detect 함수 적용.
6. NUCLEO-F411RE에 I2C2 사용할 수 있게 함.

+++ fc/include/link_driver.h
@@ -65,7 +65,7 @@ struct sensorSet_t;
 typedef struct sensorSet_t sensorSet_t;
 
 typedef BOOL (*sensorInitFuncPtr)(sensorSet_t* p_sensor_set, PifImuSensorAlign align);// sensor init prototype
-typedef BOOL (*sensorReadFuncPtr)(sensorSet_t* p_sensor_set, int16_t *data);          // sensor read and align prototype
+typedef BOOL (*sensorReadFuncPtr)(sensorSet_t* p_sensor_set, float *data);          // sensor read and align prototype
 
 struct sensorSet_t {
     PifImuSensor imu_sensor;
