+++ fc/source/imu.c
@@ -6,7 +6,7 @@
 #endif
 
 
-int16_t gyroADC[3], accADC[3], accSmooth[3], magADC[3];
+float gyroADC[3], accADC[3], accSmooth[3], magADC[3];
 int32_t accSum[3];
 uint32_t accTimeSum = 0;        // keep track for integration of acc
 int accSumCount = 0;
