+++ naze32/src/drv_adxl345.c
@@ -41,7 +41,7 @@
 extern uint16_t acc_1G;
 
 static BOOL adxl345Init(sensorSet_t *p_sensor_set, PifImuSensorAlign align);
-static BOOL adxl345Read(sensorSet_t *p_sensor_set, int16_t *accelData);
+static BOOL adxl345Read(sensorSet_t *p_sensor_set, float *accelData);
 
 static bool useFifo = false;
 static PifImuSensorAlign accAlign = IMUS_ALIGN_CW270_DEG;
@@ -93,7 +93,7 @@ static BOOL adxl345Init(sensorSet_t *p_sensor_set, PifImuSensorAlign align)
 
 uint8_t acc_samples = 0;
 
-static BOOL adxl345Read(sensorSet_t *p_sensor_set, int16_t *accelData)
+static BOOL adxl345Read(sensorSet_t *p_sensor_set, float *accelData)
 {
     uint8_t buf[8];
     int16_t data[3];
