+++ naze32/src/utils.h
@@ -8,7 +8,7 @@ extern "C" {
 __attribute__ ((weak)) int constrain(int amt, int low, int high);
 #endif
 // sensor orientation
-void alignSensors(int16_t *src, int16_t *dest, uint8_t rotation);
+void alignSensors(int16_t *src, float *dest, uint8_t rotation);
 void initBoardAlignment(void);
 
 #ifdef __cplusplus
